% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\name{calc_food_emissions}
\alias{calc_food_emissions}
\title{Calculate Food-Based Emissions using internal dataset}
\usage{
calc_food_emissions(df)
}
\arguments{
\item{df}{A data frame containing the food consumption data.}
}
\value{
A data frame with a new column \code{FoodEmissions} representing total food emissions.
}
\description{
This function computes food-based emissions by converting food intake data into estimated carbon emissions.
}