% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_housing_emissions.R
\name{calc_housing_emissions}
\alias{calc_housing_emissions}
\title{Calculate Housing Emissions}
\usage{
calc_housing_emissions(df)
}
\arguments{
\item{df}{A data frame containing housing-related data.}
}
\value{
A data frame with a new column \code{HousingEmissions} representing total housing emissions.
}
\description{
Computes housing emissions by converting energy and utility expenditure data into estimated annual carbon emissions.
}
