% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pet_emissions.R
\name{calc_pet_emissions}
\alias{calc_pet_emissions}
\title{Calculate Pet Emissions}
\usage{
calc_pet_emissions(df)
}
\arguments{
\item{df}{A data frame containing pet ownership data.}
}
\value{
A data frame with a new column \code{PetEmissions} representing total pet emissions.
}
\description{
This function computes pet-related emissions based on the number of cats and dogs owned.
}
