% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\name{calc_transport_emissions}
\alias{calc_transport_emissions}
\title{Calculate Transport Emissions}
\usage{
calc_transport_emissions(df)
}
\arguments{
\item{df}{A data frame containing transportation-related survey data.}
}
\value{
A data frame with a new column \code{TransportEmissions} representing total transport emissions.
}
\description{
This function computes transportation emissions by converting survey responses into estimated annual carbon emissions.
}