% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cons_emissions.R,
%   R/calc_cons_emissions_process.R
\name{get_cons_emission_factors}
\alias{get_cons_emission_factors}
\title{Get country-specific emission factors from the built-in dataset}
\usage{
get_cons_emission_factors(country)

get_cons_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A named list of consumption emission factors.

A named list of consumption emission factors.
}
\description{
Fetches consumption emission factors from the dataset stored in the package.

Fetches consumption emission factors from the dataset stored in the package.
}
