% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_food_emissions.R,
%   R/calc_food_emissions_process.R
\name{get_food_emission_factors}
\alias{get_food_emission_factors}
\title{Get country-specific food emission factors from the built-in dataset}
\usage{
get_food_emission_factors(country)

get_food_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A named list of food emission factors.

A named list of food emission factors.
}
\description{
Fetches food emission factors from the dataset stored in the package.

Fetches food emission factors from the dataset stored in the package.
}
