% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
%   R/calc_housing_emissions_process.R
\name{get_housing_emission_factors}
\alias{get_housing_emission_factors}
\title{Get country-specific housing emission factors from dataset}
\usage{
get_housing_emission_factors(country)

get_housing_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A list of housing emission factors.

A list of housing emission factors.
}
\description{
Returns a list of housing emission factors based on the country.

Returns a list of housing emission factors based on the country.
}