% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
%   R/calc_pet_emissions_process.R, R/get_pet_emission_factors.R
\name{get_pet_emission_factors}
\alias{get_pet_emission_factors}
\title{Get country-specific pet emission factors}
\usage{
get_pet_emission_factors(country)

get_pet_emission_factors(country)

get_pet_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A list of pet emission factors.

A list of pet emission factors.

A list of pet emission factors.
}
\description{
Returns a list of pet emission factors based on the country.

Returns a list of pet emission factors based on the country.

Returns a list of pet emission factors based on the country.
}