% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
%   R/calc_transport_emissions_process.R, R/get_transport_emission_factors.R
\name{get_transport_emission_factors}
\alias{get_transport_emission_factors}
\title{Get country-specific transport emission factors}
\usage{
get_transport_emission_factors(country)

get_transport_emission_factors(country)

get_transport_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A list of transport emission factors.

A list of transport emission factors.

A list of transport emission factors.
}
\description{
Returns a list of transport emission factors based on the country.

Returns a list of transport emission factors based on the country.

Returns a list of transport emission factors based on the country.
}