% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\docType{data}
\name{housing_emission_factors}
\alias{housing_emission_factors}
\title{Housing Emission Factors Dataset}
\format{
A data frame with columns:
\describe{
\item{Country}{Country name}
\item{FactorName}{Type of housing emission factor (e.g., Electricity, Heating_Oil, Natural_Gas, Water)}
\item{Value}{Emission value for the given factor}
}
}
\usage{
data(housing_emission_factors)
}
\description{
This dataset contains housing-related emission factors based on country-specific data.
}
\keyword{datasets}