% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\docType{data}
\name{pet_emission_factors}
\alias{pet_emission_factors}
\title{Pet Emission Factors Dataset}
\format{
A data frame with columns:
\describe{
\item{Country}{Character, name of the country or region}
\item{Dog}{Numeric, carbon emissions from dogs (kg CO2 per year)}
\item{Cat}{Numeric, carbon emissions from cats (kg CO2 per year)}
}
}
\usage{
data(pet_emission_factors)
}
\description{
This dataset contains emission factors for pet ownership based on country-specific data.
}
\keyword{datasets}