% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_data.R
\docType{data}
\name{zip_data}
\alias{zip_data}
\title{ZIP Code Data for U.S. States}
\format{
A data frame with the following columns:
\describe{
\item{ID}{Unique identifier}
\item{ST}{State abbreviation}
\item{State_Name}{Full name of the state}
\item{Zip_Max}{Maximum ZIP code in the range}
\item{Zip_Min}{Minimum ZIP code in the range}
}
}
\usage{
data(zip_data)
}
\description{
This dataset provides the ZIP code ranges for different states in the United States.
}
\keyword{datasets}
