% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\name{calc_cons_emissions}
\alias{calc_cons_emissions}
\title{Calculate Consumption-Based Emissions using internal dataset}
\usage{
calc_cons_emissions(df)
}
\arguments{
\item{df}{A data frame containing the consumption data.}
}
\value{
A data frame with a new column \code{ConsEmissions} representing total consumption emissions.
}
\description{
This function computes consumption-based emissions by converting expenditure data into estimated carbon emissions.
}