% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_food_emissions.R
\name{calc_food_emissions}
\alias{calc_food_emissions}
\title{Calculate Food-Based Emissions using internal dataset}
\usage{
calc_food_emissions(df, milk_price_rmb_per_l = 10, serving_volume_l = 0.5)
}
\arguments{
\item{df}{A data frame containing the food consumption data.}

\item{milk_price_rmb_per_l}{numeric. Milk price per litre (RMB).}

\item{serving_volume_l}{numeric. Serving volume in litres.}
}
\value{
A data frame with a new column \code{FoodEmissions} representing total food emissions.
}
\description{
This function computes food-based emissions by converting food intake data into estimated carbon emissions.
}
