% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_emission_factors.R
\docType{data}
\name{food_emission_factors}
\alias{food_emission_factors}
\title{Food Emission Factors Dataset}
\format{
A data frame with columns:
\describe{
\item{Country}{Country name}
\item{FactorName}{Type of food emission factor (e.g., Dairy, Meat, Vegan, Vegetarian)}
\item{Unit}{character}  # e.g., "kg CO2e / kg"
\item{Value}{Emission value for the given factor}
}
}
\usage{
data(food_emission_factors)
}
\description{
This dataset contains food-related emission factors based on country-specific data.
}
\keyword{datasets}
