% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_emissions.R,
%   R/calc_total_emissions_process.R
\name{get_total_emission_factors}
\alias{get_total_emission_factors}
\title{Get country-specific total emission factors}
\usage{
get_total_emission_factors(country)

get_total_emission_factors(country)
}
\arguments{
\item{country}{A character string representing the country.}
}
\value{
A list of total emission factors.

A list of total emission factors.
}
\description{
Returns a list of total emission factors based on the country.

Returns a list of total emission factors based on the country.
}
