% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-SECFC.Rmd.
\docType{data}
\name{transport_emission_factors}
\alias{transport_emission_factors}
\title{Transport Emission Factors Dataset}
\format{
A data frame with columns:
\describe{
\item{Country}{Character, name of the country or region}
\item{Diesel_Vehicles}{Numeric, carbon emissions from diesel vehicles}
\item{Electric_Vehicles}{Numeric, carbon emissions from electric vehicles}
\item{Flights}{Numeric, emissions from air travel}
\item{Gasoline_Vehicles}{Numeric, emissions from gasoline vehicles}
\item{Hybrid_Vehicles}{Numeric, emissions from hybrid vehicles}
\item{Long_Distance_Train}{Numeric, emissions from train travel}
\item{Natural_Gas_Vehicles}{Numeric, emissions from natural gas vehicles}
\item{Public_Transport}{Numeric, emissions from public transport}
}
}
\usage{
data(transport_emission_factors)
}
\description{
This dataset contains emission factors for different modes of transport,
categorized by country.
}
\keyword{datasets}